requirements.txt
ecdsa==0.13
MacFSEvents==0.4
paramiko==1.15.2
pycrypto==2.6.1
